fix(tao-windows): revert to stock Compose Desktop wheel scrolling
The Chrome-pipeline experiments (synthetic fling #309/#312/#314/#315,
DirectManipulation #313) never reached parity on-device and produced
juddery inertia. Replace the whole stack with the stock Compose Desktop
Windows behavior, matching compose-desktop-native:

- wheel events go straight into the scene; MouseWheelScrollingLogic's
  smooth tween provides the entire feel (never precise, no fling)
- TaoWindowsScrollConfig (ex ChromeScrollConfig): WindowsWinUIConfig
  formula (bounds/20 x 3 lines per notch)
- keep the inline frame pump in onPointerScroll: our frame clock is
  WM_PAINT-driven and a wheel flood starves it, freezing the tween
- remove TaoWheelFling, ChromiumFlingCurve, TaoScrollDiagnostics, the
  DirectManipulation bridge (Kotlin + C++ + smoke test + CI step +
  GraalVM metadata + build.bat), and the ScrollTestScreen pipeline labels
