+++ .github/workflows/build-natives.yaml
@@ -376,19 +376,6 @@ jobs:
             echo "OK: $f ($(wc -c < "$f") bytes)"
           done
 
-      # Exercise the freshly built DLLs against a real HWND on the Windows
-      # runner: DirectManipulation viewport smoke (COM attach, idle fetch,
-      # detach). Self-skips on non-Windows, so this is the only place it
-      # actually executes in CI. StandalonePanelNativeSmokeTest is NOT run
-      # here: its DComp-backed panel can't initialize in the runner's
-      # non-interactive session (no compositor) — it remains an on-device
-      # test only.
-      - name: Run decorated-window-tao Windows native smoke tests
-        shell: bash
-        run: |
-          ./gradlew :decorated-window-tao:test --no-daemon \
-            --tests "dev.nucleusframework.window.tao.DManipNativeSmokeTest"
-
       - name: Upload decorated-window-tao Windows DLLs
         uses: actions/upload-artifact@v4
         with:
