+++ decorated-window-tao/src/main/resources/META-INF/native-image/dev.nucleusframework/nucleus.decorated-window-tao/reachability-metadata.json
@@ -164,10 +164,6 @@
             ]
         },
         {
-            "type": "dev.nucleusframework.window.tao.NativeTaoDManipBridge",
-            "jniAccessible": true
-        },
-        {
             "type": "dev.nucleusframework.window.tao.NativeTaoGlBridge",
             "jniAccessible": true
         },
